Adiciona Domain/Entities/Carrinho.cs

+++ src/Domain/Entities/Carrinho.cs
@@ -3,7 +3,7 @@ namespace Ecommerce.Domain.Entities;
     {
         public DateTime DataCriacao { get; set; } 
         public ICollection<ItemCarrinho> Itens { get; set; } = [];
-        public double PrecoTotal { get; set; }
+        public decimal PrecoTotal { get; set; }
         public Cliente? Cliente { get; set; } 
         public int ClienteId { get; set; }
         public Pedido? Pedido { get; set; }
